+++ docs/howto/deploying-asset.md
@@ -1,0 +1,17 @@
+### Deploy Asset with every dependencies
+
+## L2
+
+note: change mumbai by the mainnet when needed
+
+```shell
+yarn deploy mumbai --tags "PolygonAssetAttributesRegistry,PolygonGemsCatalystsRegistry"
+```
+
+## L1
+
+note: change goerli by the mainnet when needed
+
+```shell
+yarn deploy goerli --tags "AssetAttributesRegistry,GemsCatalystsRegistry,AssetV2"
+```
